requirements.txt
boto==2.40.0
duo_client==2.1
elasticsearch==1.3.0
iso8601==0.1.11
python-dateutil==2.4.0
pytz==2014.10
PyYAML==3.11
requests[security]==2.12.4
simplejson==3.10.0
urllib3==1.10.3
PyStaticConfiguration
pycurl==7.19.0.2
